requirements/common.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.10.5

# Configuration
django-environ==0.4.1
whitenoise==3.2.2


# Forms
django-braces==1.10.0
django-crispy-forms==1.6.1

# Models
django-model-utils==2.6

# Images
Pillow==3.4.2

# For user registration, either via email or social
django-allauth==0.29.0


# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2016.10

# Redis support
django-redis==4.6.0
redis>=2.10.5


celery==4.0.1

rcssmin==1.0.6
django-compressor==2.1

# Admin
django-constance==1.3.3
django-admin-tools==0.8.0


# API
# django-rest-swagger==2.0.7
djangorestframework==3.5.3
django-filter==1.0.1
django-cors-headers==1.3.1
django-oauth-toolkit==0.11.0
coreapi==2.1.1
markdown==2.6.7


# OLD http://django-rest-swagger.readthedocs.io/en/stable-0.3.x/
django-rest-swagger==0.3.9


# Toosl to manage steps
transitions==0.4.2

# ~ Blog
# ------
wagtail==1.8
wagtailfontawesome==1.0.5
wagalytics==0.6
django-modelcluster==2.0
django-taggit==0.21.3
django-el-pagination==3.0.1
tapioca-disqus==0.1.2
tapioca-wrapper==0.5.3
docutils==0.12

# ~ MODELS
# ----------
# https://github.com/brack3t/django-braces
django-braces==1.10.0
# https://github.com/carljm/django-model-utils
django-model-utils==2.6
# https://github.com/django-import-export/django-import-export
django-import-export==0.5.1


# ~ SEARCH
# ----------
elasticsearch==5.0.0
# https://github.com/stephenmcd/django-overextends
django-overextends==0.4.1
